requirements.txt
torch>=1.12.0
numpy
pyyaml
scipy
open3d
